--- js/templates/create_template.js
@@ -1,7 +1,0 @@
-const createTemplate = (templateString) => {
-  const dummyWrap = document.createElement(`div`);
-  dummyWrap.innerHTML = templateString.trim();
-  return dummyWrap.firstChild;
-};
-
-export default createTemplate;
