--- js/templates/render_template.js
@@ -1,7 +1,0 @@
-const renderTemplate = (domNode) => {
-  const mainSection = document.querySelector(`.main`);
-  mainSection.innerHTML = ``;
-  mainSection.appendChild(domNode);
-};
-
-export default renderTemplate;
