+++ js/views/artist_view.js
@@ -1,17 +1,15 @@
 import AbstractView from "./abstract_view";
 import {levels} from '../data/data';
-import timerTemplate from '../templates/timer';
-
 
 const artistAnswerItem = (level, index) => `
-<div class="main-answer-wrapper">
-  <input class="main-answer-r" type="radio" id="answer-${index + 1}" name="answer" value="val-${index + 1}"/>
-  <label class="main-answer" for="answer-${index + 1}">
-    <img class="main-answer-preview" src="${level.image}"
-        alt="${level.artist}" width="134" height="134">
-    ${level.artist}
-  </label>
-</div>
+  <div class="main-answer-wrapper">
+    <input class="main-answer-r" type="radio" id="answer-${index + 1}" name="answer" value="val-${index + 1}"/>
+    <label class="main-answer" for="answer-${index + 1}">
+      <img class="main-answer-preview" src="${level.image}"
+          alt="${level.artist}" width="134" height="134">
+      ${level.artist}
+    </label>
+  </div>
 `;
 
 export default class ArtistView extends AbstractView {
@@ -22,28 +20,27 @@ export default class ArtistView extends AbstractView {
 
   get template() {
     return `
-    <section class="main main--level main--level-artist">
-      ${timerTemplate}
-      <div class="main-wrap">
-        <h2 class="title main-title">Кто исполняет эту песню?</h2>
-        <div class="player-wrapper">
-          <div class="player">
-            <audio src="${levels[this.level - 1].correctAnswerSrc}"></audio>
-            <button class="player-control"></button>
-            <div class="player-track">
-              <span class="player-status"></span>
+      <section class="main main--level main--level-artist">
+        <div class="main-wrap">
+          <h2 class="title main-title">Кто исполняет эту песню?</h2>
+          <div class="player-wrapper">
+            <div class="player">
+              <audio src="${levels[this.level - 1].correctAnswerSrc}"></audio>
+              <button class="player-control"></button>
+              <div class="player-track">
+                <span class="player-status"></span>
+              </div>
             </div>
           </div>
+          <form class="main-list">
+            ${levels[this.level - 1].artistList.map((level, index) => artistAnswerItem(level, index)).join(``)}
+          </form>
         </div>
-        <form class="main-list">
-          ${levels[this.level - 1].artistList.map((level, index) => artistAnswerItem(level, index)).join(``)}
-        </form>
-      </div>
-    </section>
-  `;
+      </section>
+    `;
   }
 
-  onAnswerClick(answerText) {
+  onAnswerClick() {
   }
 
   bind() {
